+++ lib/Credentials.php
@@ -16,12 +16,12 @@ class Credentials
      *
      * If $tokenOrFile is null, try to find a token in this order:
      *
-     * 1. Read $CATALYTIC_CREDENTIALS env var
-     * 2. Read ~/.catalytic/credentials/default
+     * 1. Read $CATALYTIC_TOKEN env var
+     * 2. Read ~/.catalytic/tokens/default
      *
      * If $tokenOrFile is not null, try to find a token in this order:
      *
-     * 1. Read ~/.catalytic/credentials/$tokenOrFile
+     * 1. Read ~/.catalytic/tokens/$tokenOrFile
      * 2. Read $tokenOrFile as the path to a file
      * 3. Assume it's an actual token that was passed in
      *
@@ -41,8 +41,8 @@ class Credentials
     /**
      * Fetch the Catalytic token.
      *
-     * First tries to fetch the token from the env var $CATALYTIC_CREDENTIALS,
-     * then tries to fetch it from ~/.catalytic/credentials/default.
+     * First tries to fetch the token from the env var $CATALYTIC_TOKEN,
+     * then tries to fetch it from ~/.catalytic/tokens/default.
      *
      * @return string   The Catalytic access token
      * @throws Exception
@@ -60,15 +60,15 @@ class Credentials
         // If it wasn't found, throw an exception
         if (!$token) {
             $home = $this->getHomeDir();
-            throw new Exception('Cannot find credentials in $CATALYTIC_CREDENTIALS
-                environment variable or ' . "$home/.catalytic/credentials/default");
+            throw new Exception('Cannot find Access Token in $CATALYTIC_TOKEN
+                environment variable or ' . "$home/.catalytic/tokens/default");
         }
 
         return $token;
     }
 
     /**
-     * Fetch the Catalytic token from a named file in ~/.catalytic/credentials/<$fileName>
+     * Fetch the Catalytic token from a named file in ~/.catalytic/tokens/<$fileName>
      *
      * @param string $fileName          The name of the file to fetch the token from
      * @return string                   The Catalytic access token
@@ -81,7 +81,7 @@ class Credentials
         // If it wasn't found, throw an exception
         if ($token === null) {
             $home = $this->getHomeDir();
-            throw new Exception('Cannot find credentials in ' . "$home/.catalytic/credentials/$fileName" . "or $fileName");
+            throw new Exception('Cannot find Access Token in ' . "$home/.catalytic/tokens/$fileName" . "or $fileName");
         }
 
         return $token;
@@ -94,7 +94,7 @@ class Credentials
      */
     private function fetchTokenFromEnvVar()
     {
-        $token = getenv('CATALYTIC_CREDENTIALS');
+        $token = getenv('CATALYTIC_TOKEN');
         return $token;
     }
 
@@ -116,9 +116,9 @@ class Credentials
 
         // If it's only the name of a file
         if ($fileName) {
-            $path = "$home/.catalytic/credentials/$fileName";
+            $path = "$home/.catalytic/tokens/$fileName";
         } else {
-            $path = "$home/.catalytic/credentials/default";
+            $path = "$home/.catalytic/tokens/default";
         }
         $token = file_get_contents($path);